Update patch set 1
Patch Set 1:

No clue, but baseband versions are a massive clusterfuck and no way anyone wants to deal with those. They look like this: 01.04_U11440601_71.02.50709G_F and the numbers vary for every release for every variant. So the developer edition has a different version that the AT&T or the European, etc. TrustZone's easier to deal with.

Patch-set: 1