Update patch set 21
Patch Set 21:

PS22 is a rebase of 21 + the hard power on patch that got abandoned, so no changes there.

I mistook the patch you were referring to though. The kernel needs this: http://review.cyanogenmod.org/#/c/117252/

That patch dumps all I2C traffic to the log, which lets you see exactly what state the TFA chip is in.

Patch-set: 21