Update patch set 6
Patch Set 6:

There are a few devices that are actually combinations of other devices ('speaker-and-headphones' being the most prominent). This patch makes the HAL select the 'speaker' device and the 'headphone' device when 'speaker-and-headphones' comes up, instead of selecting a custom 'speaker-and-headphones' device. This has the nice effect of not de-routing everything when you do switch off the custom 'speaker-and-headphones' device to just 'speaker' or just 'headphones', since the HAL knows that 'speaker-and-headphones' is actually 'speaker' and 'headphones', and that it can just turn off 'headphones' without touching 'speaker'.

Patch-set: 6