Update patch set 2
Patch Set 2:

It would be awesome (for me, and maybe others) if you could break this into 2 commits:

commit #1: import the ril-caf bits (as is, give or take any minor tweaks for directory structure, etc).

commit #2: anything that you are changing for this/these devices.

That way it is infinitely easier to see what had to be changed to make it work rather than trying to figure it out after the fact.

Patch-set: 2